+++ src/main/java/codesquad/http/HttpRequest.java
@@ -1,4 +1,4 @@
-package codesquad;
+package codesquad.http;
 
 import java.util.HashMap;
 import java.util.List;
