+++ src/main/java/codesquad/http/handler/RequestHandler.java
@@ -1,0 +1,7 @@
+package codesquad.http.handler;
+
+import java.net.Socket;
+
+public interface RequestHandler {
+    void handleRequest(Socket clientSocket);
+}
